refactor(mail): Implement Strategy pattern for MailService
Refactored the MailService to use the Strategy pattern, addressing reviewer feedback to create a more modular and testable design. The mail sending logic is now split into two distinct strategies: the legacy App Engine API and the new SMTP implementation.

Key changes:
- Renamed the original  to  to clarify that it uses the traditional App Engine API.
- Created a new  to encapsulate all logic for sending email via an external SMTP server.
- Updated  to act as a factory. It now checks the  environment variable to determine which  implementation ( or ) to provide.
- Reorganized the tests to match the new structure:
    - Renamed  to  to focus exclusively on the SMTP implementation.
    - Created  to verify the factory's selection logic.

This new structure improves adherence to the Single Responsibility and Open/Closed principles, making the system more scalable and easier to maintain.
