+++ api/src/main/java/com/google/appengine/api/mail/SmtpMailServiceImpl.java
@@ -16,12 +16,6 @@
 
 package com.google.appengine.api.mail;
 
-import com.google.appengine.api.mail.MailServicePb.MailAttachment;
-import com.google.appengine.api.mail.MailServicePb.MailHeader;
-import com.google.appengine.api.mail.MailServicePb.MailMessage;
-import com.google.appengine.api.mail.MailServicePb.MailServiceError.ErrorCode;
-import com.google.apphosting.api.ApiProxy;
-import com.google.protobuf.ByteString;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -31,10 +25,8 @@ import java.util.Properties;
 import javax.activation.DataHandler;
 import javax.activation.DataSource;
 import javax.mail.Address;
-import javax.mail.BodyPart;
 import javax.mail.Message.RecipientType;
 import javax.mail.MessagingException;
-import javax.mail.Multipart;
 import javax.mail.Session;
 import javax.mail.Transport;
 import javax.mail.internet.AddressException;
@@ -45,38 +37,27 @@ import javax.mail.internet.MimeMultipart;
 import javax.mail.util.ByteArrayDataSource;
 
 /**
- * This class implements raw access to the mail service.
- * Applications that don't want to make use of Sun's JavaMail
- * can use it directly -- but they will forego the typing and
- * convenience methods that JavaMail provides.
- *
+ * This class implements the MailService interface using an external SMTP server.
  */
-class MailServiceImpl implements MailService {
-  static final String PACKAGE = "mail";
+class SmtpMailServiceImpl implements MailService {
   private final EnvironmentProvider envProvider;
 
-  /** Default constructor, used in production. */
-  MailServiceImpl() {
-    this(new SystemEnvironmentProvider());
-  }
-
-  /** Constructor for testing, allowing a mock environment provider. */
-  MailServiceImpl(EnvironmentProvider envProvider) {
+  /**
+   * Constructor.
+   * @param envProvider The provider for environment variables.
+   */
+  SmtpMailServiceImpl(EnvironmentProvider envProvider) {
     this.envProvider = envProvider;
   }
 
-  /** {@inheritDoc} */
   @Override
-  public void sendToAdmins(Message message)
-      throws IllegalArgumentException, IOException {
-    doSend(message, true);
+  public void send(Message message) throws IOException {
+    sendSmtp(message, false);
   }
 
-  /** {@inheritDoc} */
   @Override
-  public void send(Message message)
-      throws IllegalArgumentException, IOException {
-    doSend(message, false);
+  public void sendToAdmins(Message message) throws IOException {
+    sendSmtp(message, true);
   }
 
   private void sendSmtp(Message message, boolean toAdmin)
@@ -141,7 +122,6 @@ class MailServiceImpl implements MailService {
       if (!ccRecipients.isEmpty()) {
         mimeMessage.setRecipients(RecipientType.CC, ccRecipients.toArray(new Address[0]));
       }
-      // Bcc recipients are not set on the MimeMessage to prevent them from being in the headers.
 
       if (message.getReplyTo() != null) {
         mimeMessage.setReplyTo(new Address[] {new InternetAddress(message.getReplyTo())});
@@ -154,14 +134,11 @@ class MailServiceImpl implements MailService {
       final boolean hasAmpHtmlBody = message.getAmpHtmlBody() != null;
       final boolean hasTextBody = message.getTextBody() != null;
 
-      // Case 1: Plain text only, no attachments. Simplest case.
       if (hasTextBody && !hasHtmlBody && !hasAmpHtmlBody && !hasAttachments) {
         mimeMessage.setText(message.getTextBody());
       } else {
-        // Case 2: Anything more complex requires multipart.
         MimeMultipart topLevelMultipart = new MimeMultipart("mixed");
 
-        // The bodies (text, html, amp) are grouped in a "multipart/alternative"
         if (hasTextBody || hasHtmlBody || hasAmpHtmlBody) {
             MimeMultipart alternativeMultipart = new MimeMultipart("alternative");
             MimeBodyPart alternativeBodyPart = new MimeBodyPart();
@@ -172,7 +149,6 @@ class MailServiceImpl implements MailService {
                 textPart.setText(message.getTextBody());
                 alternativeMultipart.addBodyPart(textPart);
             } else if (hasHtmlBody) {
-                // If there is an HTML body but no text body, add an empty text part for compatibility.
                 MimeBodyPart textPart = new MimeBodyPart();
                 textPart.setText("");
                 alternativeMultipart.addBodyPart(textPart);
@@ -191,7 +167,6 @@ class MailServiceImpl implements MailService {
             topLevelMultipart.addBodyPart(alternativeBodyPart);
         }
 
-        // Add attachments to the top-level mixed part.
         if (hasAttachments) {
           for (Attachment attachment : message.getAttachments()) {
             MimeBodyPart attachmentBodyPart = new MimeBodyPart();
@@ -214,7 +189,6 @@ class MailServiceImpl implements MailService {
         }
       }
 
-      // Update headers to match content, e.g., setting the Content-Type
       mimeMessage.saveChanges();
 
       Transport transport = session.getTransport("smtp");
@@ -242,100 +216,5 @@ class MailServiceImpl implements MailService {
       list.add(new InternetAddress(address));
     }
     return list;
-  }
-
-  /**
-   * Does the actual sending of the message.
-   * @param message The message to be sent.
-   * @param toAdmin Whether the message is to be sent to the admins.
-   */
-  private void doSend(Message message, boolean toAdmin)
-      throws IllegalArgumentException, IOException {
-    if ("true".equals(envProvider.getenv("APPENGINE_USE_SMTP_MAIL_SERVICE"))) {
-      sendSmtp(message, toAdmin);
-      return;
-    }
-    // Could perform basic checks to save on RPCs in case of missing args etc.
-    // I'm not doing this on purpose, to make sure the semantics of the two
-    // implementations stay the same.
-    // The benefit of not doing basic checks here is that we will pick up
-    // changes in semantics (ie from address can now also be the logged-in user)
-    // for free.
-
-    MailMessage.Builder msgProto = MailMessage.newBuilder();
-    if (message.getSender() != null) {
-      msgProto.setSender(message.getSender());
-    }
-    if (message.getTo() != null) {
-      msgProto.addAllTo(message.getTo());
-    }
-    if (message.getCc() != null) {
-      msgProto.addAllCc(message.getCc());
-    }
-    if (message.getBcc() != null) {
-      msgProto.addAllBcc(message.getBcc());
-    }
-    if (message.getReplyTo() != null) {
-      msgProto.setReplyTo(message.getReplyTo());
-    }
-    if (message.getSubject() != null) {
-      msgProto.setSubject(message.getSubject());
-    }
-    if (message.getTextBody() != null) {
-      msgProto.setTextBody(message.getTextBody());
-    }
-    if (message.getHtmlBody() != null) {
-      msgProto.setHtmlBody(message.getHtmlBody());
-    }
-    if (message.getAmpHtmlBody() != null) {
-      msgProto.setAmpHtmlBody(message.getAmpHtmlBody());
-    }
-    if (message.getAttachments() != null) {
-      for (Attachment attach : message.getAttachments()) {
-        MailAttachment.Builder attachProto = MailAttachment.newBuilder();
-        attachProto.setFileName(attach.getFileName());
-        attachProto.setData(ByteString.copyFrom(attach.getData()));
-        String contentId = attach.getContentID();
-        if (contentId != null) {
-          attachProto.setContentID(contentId);
-        }
-        msgProto.addAttachment(attachProto);
-      }
-    }
-    if (message.getHeaders() != null) {
-      for (Header header : message.getHeaders()) {
-        msgProto.addHeader(
-            MailHeader.newBuilder().setName(header.getName()).setValue(header.getValue()));
-      }
-    }
-
-    byte[] msgBytes = msgProto.buildPartial().toByteArray();
-    try {
-      // Returns VoidProto -- just ignore the return value.
-      if (toAdmin) {
-        ApiProxy.makeSyncCall(PACKAGE, "SendToAdmins", msgBytes);
-      } else {
-        ApiProxy.makeSyncCall(PACKAGE, "Send", msgBytes);
-      }
-    } catch (ApiProxy.ApplicationException ex) {
-      // Pass all the error details straight through (same as python).
-      switch (ErrorCode.forNumber(ex.getApplicationError())) {
-        case BAD_REQUEST:
-          throw new IllegalArgumentException("Bad Request: " +
-                                             ex.getErrorDetail());
-        case UNAUTHORIZED_SENDER:
-          throw new IllegalArgumentException("Unauthorized Sender: " +
-                                             ex.getErrorDetail());
-        case INVALID_ATTACHMENT_TYPE:
-          throw new IllegalArgumentException("Invalid Attachment Type: " +
-                                             ex.getErrorDetail());
-        case INVALID_HEADER_NAME:
-          throw new IllegalArgumentException("Invalid Header Name: " +
-                                             ex.getErrorDetail());
-        case INTERNAL_ERROR:
-        default:
-          throw new IOException(ex.getErrorDetail());
-      }
-    }
   }
 }
